+++ src/Repository/ShopsRepository.php
@@ -4,6 +4,8 @@ namespace Nokaut\ApiKit\Repository;
 
 
 use Nokaut\ApiKit\ClientApi\Rest\Fetch\ShopsFetch;
+use Nokaut\ApiKit\ClientApi\Rest\Query\Filter\MultipleWithOperator;
+use Nokaut\ApiKit\ClientApi\Rest\Query\Filter\Single;
 use Nokaut\ApiKit\ClientApi\Rest\Query\Filter\SingleWithOperator;
 use Nokaut\ApiKit\ClientApi\Rest\Query\ShopsQuery;
 use Nokaut\ApiKit\Collection\Shops;
@@ -23,14 +25,51 @@ class ShopsRepository extends RepositoryAbstract
      */
     public function fetchByNamePrefix($namePrefix, $limit)
     {
-        $query = new ShopsQuery($this->apiBaseUrl);
-        $query->setFields(self::$fieldsAll);
-        $query->addFilter(new SingleWithOperator('name', 'prefix', $namePrefix));
-        $query->setLimit($limit);
+        $query = $this->prepareQueryByNamePrefix($namePrefix, $limit);
 
         $fetch = new ShopsFetch($query, $this->cache);
         $this->clientApi->send($fetch);
 
         return $fetch->getResult();
+    }
+
+    /**
+     * @param array $ids
+     * @return Shops
+     */
+    public function fetchByIds(array $ids)
+    {
+        $query = $this->prepareQueryByIds($ids);
+
+        $fetch = new ShopsFetch($query, $this->cache);
+        $this->clientApi->send($fetch);
+
+        return $fetch->getResult();
+    }
+
+    /**
+     * @param $namePrefix
+     * @param $limit
+     * @return ShopsQuery
+     */
+    protected function prepareQueryByNamePrefix($namePrefix, $limit)
+    {
+        $query = new ShopsQuery($this->apiBaseUrl);
+        $query->setFields(self::$fieldsAll);
+        $query->addFilter(new SingleWithOperator('name', 'prefix', $namePrefix));
+        $query->setLimit($limit);
+        return $query;
+    }
+
+    /**
+     * @param array $ids
+     * @return ShopsQuery
+     */
+    protected function prepareQueryByIds(array $ids)
+    {
+        $query = new ShopsQuery($this->apiBaseUrl);
+        $query->setFields(self::$fieldsAll);
+        $query->addFilter(new MultipleWithOperator('id', 'in', $ids));
+        return $query;
     }
 }