+++ CHANGELOG.md
@@ -4,7 +4,9 @@ ChangeLog
 master
 -----
 - Dodanie atrybutu is_label do cechy produktu
-- zmiana interface \Nokaut\ApiKit\Cache\CacheInterface oraz dodanie Nokaut\ApiKit\Cache\AbstractCache - jeśli implementowaliście interface CacheInterface wystarczy, że rozszerzycie swoją klasę abstrkcyjną klasą Nokaut\ApiKit\Cache\AbstractCache 
+- zmiana interface \Nokaut\ApiKit\Cache\CacheInterface oraz dodanie Nokaut\ApiKit\Cache\AbstractCache - jeśli implementowaliście interface CacheInterface wystarczy, że rozszerzycie swoją klasę abstrkcyjną klasą Nokaut\ApiKit\Cache\AbstractCache
+- Pobieranie ShopsRepository i ProducersRepository z klasy ApiKit
+- Dodanie repozytoriów: ShopsAsyncRepository i ProducersAsyncRepository
 
 v1.6.0
 ------